fix safari bug. fix #57
Switch to call all functions of the Client struct on a copy of the inner WgClient struct to prevent recursive access due to async functions/ callbacks.

Create the js object returned in the body of the WgClient::fetch before creating the Respone to workaround a bug in wasm_bindgen: https://github.com/rustwasm/wasm-bindgen/issues/2648

Prepare use of log crate.